+++ Flow/src/utopia/flow/collection/immutable/caching/cache/Cache.scala
@@ -78,7 +78,7 @@ object Cache
 	{
 		// ATTRIBUTES	------------------
 		
-		private val cachedItems: mutable.Map[Key, Value] = mutable.Map()
+		private val cachedItems: mutable.Map[Any, Value] = mutable.Map()
 		
 		
 		// IMPLEMENTED	------------------