+++ Vault/src/utopia/vault/database/DatabaseCache.scala
@@ -80,7 +80,6 @@ class DatabaseCache[A, Key](connectionPool: ConnectionPool, accessor: SingleMode
 	override def cachedValues = cache.cachedValues
 	
 	override def apply(key: Key) = cache(key)
-	
 	override def cached(key: Key) = cache.cached(key)